[#705] Clean-up _class8 utility modules (_compute => _compute + _move + _los_data)
